minor changes & started moving content data to separate files
