+++ package.json
@@ -17,7 +17,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
     "react-router": "^6.22.0",
-    "react-router-dom": "^6.22.0",
+    "react-router-dom": "^6.30.2",
     "styled-components": "^6.0.7"
   },
   "devDependencies": {
