+++ src/index.css
@@ -26,6 +26,8 @@
   --surface-container-text: #e7e7e7;
   --surface-2-container-text-active: #c2ffd3;
 
+  --surface-mobile-header: #0c0b0f;
+
   --display-xl-hero-font-size: 112px;
   --display-xl-font-size: 88px;
   --display-l-font-size: 57px;
